+++ exec.c
@@ -35,7 +35,7 @@ exec(char *path, char **argv)
   // Load program into memory.
   sz = 0;
   for(i=0, off=elf.phoff; i<elf.phnum; i++, off+=sizeof(ph)){
-    if(readi(ip, (char*)&ph, off, sizeof(ph)) != sizeof(ph))
+    if(readi(ip, (char*)&ph, off, sizeof(ph)) != (int)sizeof(ph))
       goto bad;
     if(ph.type != ELF_PROG_LOAD)
       continue;
